feat(kanban): browser nodes on the board — cards + live webview in the card modal
Browser nodes now appear as board cards (web chip) and open a live Chromium webview in the card
modal (address bar + back/fwd/reload), navigable like the canvas node. Extracted BrowserNode's
webview/toolbar into a shared BrowserSurface used by both surfaces; the two webviews for one node
stay in sync (navigation persists to data.url and re-seeds the other). '+ New → Browser' creates
one in a column. Server Edition degrades gracefully (webview inert, as before).
